LayoutTests/fast/canvas/webgl/gl-get-calls.html

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css"/>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<canvas id="canvas" width="2" height="2"> </canvas>

<script>

description("This test ensures basic functionality of the underlying graphics library");

debug("");
debug("Canvas.getContext");

var context = document.getElementById("canvas").getContext("experimental-webgl");
if (!context)
    testFailed("context does not exist");
else {
    testPassed("context exists");

    debug("");
    debug("Context is correct type");
    if (context instanceof WebGLRenderingContext)
      testPassed("context is correct type");
    else
      testFailed("context is not correct type");
  
    debug("");
    debug("Context contains getError");
    if ("getError" in context)
      testPassed("context contains getError");
    else
      testFailed("context does not contains getError");

    debug("");
    debug("Check default values");
    shouldBe('context.getParameter(context.ACTIVE_TEXTURE)', 'context.TEXTURE0');
    shouldBe('(context.getParameter(context.ALIASED_LINE_WIDTH_RANGE)[0] == 1) || (context.getParameter(context.ALIASED_LINE_WIDTH_RANGE)[1] == 1)', 'true');
    shouldBe('(context.getParameter(context.ALIASED_POINT_SIZE_RANGE)[0] == 1) || (context.getParameter(context.ALIASED_POINT_SIZE_RANGE)[1] == 1)', 'true');
    shouldBe('context.getParameter(context.ARRAY_BUFFER_BINDING)', 'null');
    shouldBe('context.getParameter(context.BLEND)', 'false');
    shouldBe('context.getParameter(context.BLEND_COLOR)', '[0, 0, 0, 0]');
    shouldBe('context.getParameter(context.BLEND_DST_ALPHA)', '0');
    shouldBe('context.getParameter(context.BLEND_DST_RGB)', '0');
    shouldBe('context.getParameter(context.BLEND_EQUATION_ALPHA)', 'context.FUNC_ADD');
    shouldBe('context.getParameter(context.BLEND_EQUATION_RGB)', 'context.FUNC_ADD');
    shouldBe('context.getParameter(context.BLEND_SRC_ALPHA)', '1');
    shouldBe('context.getParameter(context.BLEND_SRC_RGB)', '1');
    shouldBe('context.getParameter(context.COLOR_CLEAR_VALUE)', '[0, 0, 0, 0]');
    shouldBe('context.getParameter(context.COLOR_WRITEMASK)', '[1, 1, 1, 1]');
    shouldBe('context.getParameter(context.CULL_FACE)', 'false');
    shouldBe('context.getParameter(context.CULL_FACE_MODE)', 'context.BACK');
    shouldBe('context.getParameter(context.CURRENT_PROGRAM)', 'null');
    shouldBe('context.getParameter(context.DEPTH_FUNC)', 'context.LESS');
    shouldBe('context.getParameter(context.DEPTH_RANGE)', '[0, 1]');
    shouldBe('context.getParameter(context.DEPTH_TEST)', 'false');
    shouldBe('context.getParameter(context.DEPTH_WRITEMASK)', 'true');
    shouldBe('context.getParameter(context.DITHER)', 'true');
    shouldBe('context.getParameter(context.ELEMENT_ARRAY_BUFFER_BINDING)', 'null');
    shouldBe('context.getParameter(context.FRONT_FACE)', 'context.CCW');
    shouldBe('context.getParameter(context.GENERATE_MIPMAP_HINT)', 'context.DONT_CARE');
    shouldBe('context.getParameter(context.LINE_WIDTH)', '1');
    shouldBe('context.getParameter(context.POLYGON_OFFSET_FACTOR)', '0');
    shouldBe('context.getParameter(context.POLYGON_OFFSET_FILL)', 'false');
    shouldBe('context.getParameter(context.POLYGON_OFFSET_UNITS)', '0');
    shouldBe('context.getParameter(context.RENDERBUFFER_BINDING)', 'null');
    shouldBe('context.getParameter(context.SAMPLE_COVERAGE_INVERT)', 'false');
    shouldBe('context.getParameter(context.SAMPLE_COVERAGE_VALUE)', '1');
    shouldBe('context.getParameter(context.SCISSOR_BOX)[0]', '0');
    shouldBe('context.getParameter(context.SCISSOR_BOX)[1]', '0');
    shouldBe('context.getParameter(context.SCISSOR_TEST)', 'false');
    shouldBe('context.getParameter(context.STENCIL_BACK_FAIL)', 'context.KEEP');
    shouldBe('context.getParameter(context.STENCIL_BACK_FUNC)', 'context.ALWAYS');
    shouldBe('context.getParameter(context.STENCIL_BACK_PASS_DEPTH_FAIL)', 'context.KEEP');
    shouldBe('context.getParameter(context.STENCIL_BACK_PASS_DEPTH_PASS)', 'context.KEEP');
    shouldBe('context.getParameter(context.STENCIL_BACK_REF)', '0');
    
    // FIXME: Current WebKit implementation returns the wrong value (https://bugs.webkit.org/show_bug.cgi?id=31842)
    //shouldBe('context.getParameter(context.STENCIL_BACK_VALUE_MASK)', '0xFFFFFFFF');
    //shouldBe('context.getParameter(context.STENCIL_BACK_WRITEMASK)', '0xFFFFFFFF');
    // If EXT_packed_depth_stencil is supported, STENCIL_BITS > 0; otherwise, STENCIL_BITS == 0.
    shouldBe('context.getParameter(context.STENCIL_BITS) >= 0', 'true');
    shouldBe('context.getParameter(context.STENCIL_CLEAR_VALUE)', '0');
    shouldBe('context.getParameter(context.STENCIL_FAIL)', 'context.KEEP');
    shouldBe('context.getParameter(context.STENCIL_FUNC)', 'context.ALWAYS');
    shouldBe('context.getParameter(context.STENCIL_PASS_DEPTH_FAIL)', 'context.KEEP');
    shouldBe('context.getParameter(context.STENCIL_PASS_DEPTH_PASS)', 'context.KEEP');
    shouldBe('context.getParameter(context.STENCIL_REF)', '0');
    shouldBe('context.getParameter(context.STENCIL_TEST)', 'false');
    //shouldBe('context.getParameter(context.STENCIL_VALUE_MASK)', '0xFFFFFFFF');
    //shouldBe('context.getParameter(context.STENCIL_WRITEMASK)', '0xFFFFFFFF');
    shouldBe('context.getParameter(context.TEXTURE_BINDING_2D)', 'null');
    shouldBe('context.getParameter(context.TEXTURE_BINDING_CUBE_MAP)', 'null');
    shouldBe('context.getParameter(context.VIEWPORT)', '[0, 0, 2, 2]');
}

debug("");
successfullyParsed = true;

</script>
<script src="../../js/resources/js-test-post.js"></script>

<script>
</script>

</body>
</html>
